xlog_cuda

Module cuda_compat

Re-exports

pub use crate::device::CudaFunction;

Modules

sys

Structs

CudaSlice

Vec<T> on a cuda device. You can allocate and modify this with CudaStream.

CudaStream

A wrapper around sys::CUstream that you can schedule work on.

CudaView

&[T] on a cuda device. An immutable sub-view into a CudaSlice created by CudaSlice::as_view()/CudaSlice::slice().

CudaViewMut

&mut [T] on a cuda device. A mutable sub-view into a CudaSlice created by CudaSlice::as_view_mut()/CudaSlice::slice_mut().

DeviceParamStorage

DriverError

Wrapper around sys::CUresult. See nvidia’s CUresult docs

LaunchConfig

Configuration for [result::launch_kernel]

ScalarParamStorage

Traits

AsKernelParam

Backwards-compatible as_kernel_param() helper for manual raw launch lists.

DevicePtr

Abstraction over CudaSlice/CudaView

DevicePtrMut

Abstraction over CudaSlice/CudaViewMut

DeviceRepr

Something that can be copied to device memory and turned into a parameter for [result::launch_kernel].

DeviceSlice

Base trait for abstracting over CudaSlice/CudaView/CudaViewMut.

IntoKernelParamStorage

Convert a launch argument into storage that lives until cuLaunchKernel runs.

KernelParamStorage

Stable host-side storage for a kernel argument pointer.

KernelScalar

Scalar kernel parameters that can be copied directly into launch storage.

LaunchAsync

Old cudarc-style launch trait reimplemented on top of CUDA 13-compatible raw kernel launches.

ValidAsZeroBits

Marker trait to indicate that the type is valid when all of its bits are set to 0.