xlog_cuda::cuda_compat

Trait DeviceRepr

pub unsafe trait DeviceRepr { }

Something that can be copied to device memory and turned into a parameter for [result::launch_kernel].

Safety

This is unsafe because a struct should likely be #[repr(C)] to be represented in cuda memory, and not all types are valid.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".