xlog_cuda

Struct CudaView

pub struct CudaView<'a, T> { /* private fields */ }

&[T] on a cuda device. An immutable sub-view into a CudaSlice created by CudaSlice::as_view()/CudaSlice::slice().

Implementations

impl<T> CudaView<'_, T>

pub fn len(&self) -> usize

The number of elements T in this view.

pub fn is_empty(&self) -> bool

impl<'a, T> CudaView<'a, T>

pub fn view_ptr(self, stream: &'a CudaStream) -> (u64, SyncOnDrop<'a>)

Identical behavior to DevicePtr::device_ptr(), but the lifetime on the returned [SyncOnDrop], matches the lifetime of the view.

impl<'a, T> CudaView<'a, T>

pub fn slice(&self, bounds: impl RangeBounds<usize>) -> CudaView<'a, T>

Creates a CudaView at the specified offset from the start of self.

Panics if range.start >= self.len.

Example

let mut slice = stream.alloc_zeros::<u8>(100).unwrap();
let mut view = slice.slice(0..50);
let mut view2 = view.slice(0..25);
do_something(&view);

pub fn try_slice( &self, bounds: impl RangeBounds<usize>, ) -> Option<CudaView<'a, T>>

Fallible version of CudaView::slice

pub unsafe fn transmute<S>(&self, len: usize) -> Option<CudaView<'a, S>>

Reinterprets the slice of memory into a different type. len is the number of elements of the new type S that are expected. If not enough bytes are allocated in self for the view, then this returns None.

Safety

This is unsafe because not the memory for the view may not be a valid interpretation for the type S.

pub fn split_at(&self, mid: usize) -> (CudaView<'a, T>, CudaView<'a, T>)

pub fn try_split_at( &self, mid: usize, ) -> Option<(CudaView<'a, T>, CudaView<'a, T>)>

Fallible version of CudaSlice::split_at.

Returns None if mid > self.len.

pub fn chunks_exact( &self, chunk_size: usize, ) -> impl Iterator<Item = CudaView<'a, T>>

Returns an iterarow over subviews of size chunk_size. Differs from std::slice::ChunksExact, in that it asserts that the chunk_size must divide evenly into the length, instead of returning a remainder.